x11/libxcvt: add new port
libxcvt is a library providing a standalone version of the X server
implementation of the VESA CVT standard timing modelines generator.

libxcvt also provides a standalone version of the command line tool
cvt copied from the Xorg implementation and is meant to be a direct
replacement to the version provided by the Xorg server.

https://gitlab.freedesktop.org/ofourdan/libxcvt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@569219 35697150-7ecd-e111-bb59-0022644237b5