Add support for USES=cmake:testing
CMake-based ports have a "standard" way of controlling whether
testing should be built, by passing -DBUILD_TESTING=ON at the
configure stage (with some footnotes). Add a :testing modifier
for USES=cmake that enables a boilerplate do-test target that
rebuilds with testing enabled, and then runs the tests.

Individual ports need to buy in to this explicitly (because
tests might not be non-destructive).

Submitted and explained well by yuri@

PR:		249024
Submitted by:	yuri


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@569551 35697150-7ecd-e111-bb59-0022644237b5